docs/pbook.py: correct faulty check for PDF output
Frank K[_\c3_][_\bc_]ster discovered that docs/pbook.py produces buggy
LaTeX code, a faulty check for PDF output.  It doesn't show up if you
only try to create PDF.

But if you want to go the traditional route and produce dvi files (and
from that PostScript and maybe finally again PDF), this will fail and
still produce PDF.  The script itself doesn't do that (or rather, it
does that in fact but ignores the output in this first run), but it's
still an uglyness.

Moreover, it's completely useless, because the only package that uses
the information, graphicx, will detect the output format itself (and
in a proper way).

  http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=390629