eval-when special form
Added the EVAL-WHEN special form to the Parenscript language.  It
works similarly to the EVAL-WHEN form in Lisp.  It has the following
form:

(eval-when body-forms-language? (situation*) body-forms*)

SITUATION is either :compile-toplevel or :execute. 

BODY-FORMS-LANGUAGE is optional and either :parenscript or :lisp.  It
defaults to :lisp when :COMPILE-TOPLEVEL is specified and :parenscript
when :EXECUTE is specified.

Parenscript's EVAL-WHEN is relevant when loading toplevel forms,
either from Parenscript files or from within Lisp.