ASDF interaction for Parenscript files
see the new paren-asdf.lisp and updated compilation-interface.lisp
files for implementation details.  The function COMPILE-SCRIPT-SYSTEM
compiles all the system's files and outputs them to a single stream.
Currently it compiles each file separately; in the future we will
probably need to compile the whole program and then translate it.