defscriptmacro single string body fix
A js macro that evaluates to a string literal can't be used:

  (defjsmacro blah () "abc")
  (js (blah)) =>
     "null;"

This seems wrong. For one thing, other literals work fine:

  (defjsmacro blah () 123)
  (js (blah)) =>
     "123"

For another, Lisp behaves differently:

  (defmacro blah () "abc")
  (blah) =>
      "abc"

Actually, I'd rather use a symbol macro but it appears there is no
define-symbol-macro in Parenscript and in this case symbol-macrolet is
not convenient.

Signed-off-by: Red Daly <reddaly@gmail.com>