Added {} to *ps-literals*.
Using {} as a literal to create an empty object literal was only
working by accident previously.  When not in the function position, {}
was being compiled as a JS variable rather than an object-literal,
which just happened to work sometimes.  In concert with correctly
checking symbols that aren't in the parenscript package, this should
provide for more reliable behavior.