Added support for getters and setters in object literals.
Getters and setter are defined in ECMAScript 5.1 sec. 11.1.5.

A getter would be specified in a (CREATE ...) form as a name-value
pair of the form (GET <NAME>) <EXPR>, where <NAME> is a symbol naming
the object slot, and <EXPR> is the body of the getter.

A setter would be specified as a name-value pair of the form
(SET <NAME> <NEW-PARAM>) <EXPR>, where <NAME> is a symbol naming
the object slot, <NEW-PARAM> is a symbol naming the setter's formal
parameter for the assignment value, and <EXPR> is the body of the
setter.