Undid 88c61599074dfaa40fae285f62f8910e0a201561
No longer going to parenthesize lambdas and objects at toplevel.
The REPL use case (the original motivation for parenthesizing at
toplevel[1]) is a not very compelling special case. Apparently,
MongoDB has some brain damage with parsing parenthesized expressions,
which is a more compelling use case, and a reason not to parenthesize.

Thanks to Lukas Phaf for the bug report.

[1] https://mailman.common-lisp.net/pipermail/parenscript-devel/2011-November/000838.html